remove auto-center and scrolling
- new folders for Results and Stakeholder
- isMobile is now a helper variable
- some improved scrolling in mobiel view on lists
- remove autocentering after clicking on every map pin on mobile
- fixed back button position
