+++ client/src/helpers/isMobile.js
@@ -1,0 +1,5 @@
+const isMobile = new RegExp("Mobi", "i").test(navigator.userAgent)
+  ? true
+  : false;
+
+export default isMobile;
